feat(tui): Nord theme + bottom_pane split + AppCommand decoupling
- theme: Nord16 palette base colors + 4 semantic token groups
  (UI surface, Diff, Markdown, Syntax) matching OpenCode style
- bottom_pane: split 656-line file into mod.rs / composer.rs / status.rs
  each sub-400 lines, matching Codex bottom_pane/ directory shape
- app_command: new AppCommand enum (12 variants) with
  commands_for_event() pure-mapping function, decoupling
  UI events from agent operations per Codex pattern
- event_loop: call commands_for_event before dispatch as a planning hook
- all imports updated; cargo check clean
