+++ src/tui/render/bottom_pane_tests.rs
@@ -4,10 +4,10 @@ use ratatui::{layout::Rect, style::Color, text::Line};
 use tempfile::tempdir;
 use tokio::sync::mpsc;
 
-use super::{
-    activity_status_line, animated_activity_label, composer_hint, composer_hint_line,
-    footer_summary_text, wrapped_text_cursor_position, wrapped_text_rows,
+use super::bottom_pane::composer::{
+    composer_hint, composer_hint_line, wrapped_text_cursor_position, wrapped_text_rows,
 };
+use super::bottom_pane::status::{activity_status_line, animated_activity_label, footer_summary_text};
 use crate::config::ConfigManager;
 use crate::tui::state::{
     InteractionKind, PendingInteractionSnapshot, RunningTask, RuntimePhase, RuntimeSnapshot,
